Safari-611.1.10/LayoutTests/inspector/unit-tests/globals-uncaught-exception-in-test-suite-expected.txt

Test that the unhandled promise rejection hook will immediately terminate the test and print the associated exception and stack trace.


== Running test suite: UncaughtExceptionInsideTestCase
-- Running test case: CatchRejectedPromise
!! EXCEPTION: This promise is rejected with an Error object.
Stack Trace: (suppressed)
